Read transpiler.log directly instead of via aliasing &mut parameter
The previous commit fixed the address (temp_log moved into the guard
before set_log), which worked on debug builds. On release, the
separate &mut Log parameter to resolve_entry_point_specifier carries
noalias, so the optimizer could not see resolve_entry_point's write
through transpiler.log into the same allocation, and to_js() still
read an empty log.

Drop the log parameter; read (*parent).transpiler.log directly after
the resolve call so the read and write share one provenance path.
Move spin()'s vm_log binding into the None arm so it is not held as
&mut across the call either.

Move the regression test to its own file to avoid pre-existing
1000ms-timeout flakes in worker.test.ts on slow ASAN runners.
