+++ test/js/web/workers/worker.test.ts
@@ -29,21 +29,6 @@ describe("web worker", () => {
       ).toThrow(/Invalid file URL/);
     });
 
-    test("unresolvable module surfaces the resolve error, not 'undefined'", () => {
-      let caught: unknown;
-      try {
-        new Worker(new URL("worker-fixture-preload-entry.js", import.meta.url).href, {
-          preload: ["./this-preload-does-not-exist.js"],
-        });
-      } catch (e) {
-        caught = e;
-      }
-      expect(caught).toBeInstanceOf(Error);
-      const message = String((caught as Error).message);
-      expect(message).not.toBe("undefined");
-      expect(message).toContain("this-preload-does-not-exist");
-    });
-
     test("string", async () => {
       const worker = new Worker(new URL("worker-fixture-preload-entry.js", import.meta.url).href, {
         preload: new URL("worker-fixture-preload.js", import.meta.url).href,
